Autocode/Autocode/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "datamodel-profiler"
dynamic = ["version"]
description = "A production-ready Python package for data profiling and schema inference"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "DataModel Profiler Team", email = "team@datamodelprofiler.com"},
]
maintainers = [
    {name = "DataModel Profiler Team", email = "team@datamodelprofiler.com"},
]
keywords = [
    "data-profiling",
    "schema-inference", 
    "data-quality",
    "data-analysis",
    "pandas",
    "polars",
    "great-expectations"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "pandas>=1.5.0",
    "pyarrow>=10.0.0",
    "numpy>=1.21.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "jsonschema>=4.0.0",
    "avro-python3>=1.10.0",
    "openpyxl>=3.0.0",
    "PyYAML>=6.0.0",
    "great-expectations>=0.17.0",
]

[project.optional-dependencies]
html = ["ydata-profiling>=4.0.0"]
polars = ["polars>=0.20.0"]
orc = ["pyarrow>=10.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/your-org/datamodel_profiler"
Documentation = "https://datamodelprofiler.readthedocs.io"
Repository = "https://github.com/your-org/datamodel_profiler.git"
"Bug Tracker" = "https://github.com/your-org/datamodel_profiler/issues"

[project.scripts]
datamodel-profiler = "datamodel_profiler.cli:main"

[tool.setuptools]
packages = ["datamodel_profiler"]

[tool.setuptools.package-data]
datamodel_profiler = ["py.typed"]

[tool.setuptools_scm]
write_to = "datamodel_profiler/_version.py"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "pyarrow.*",
    "avro.*",
    "openpyxl.*",
    "yaml.*",
    "great_expectations.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=datamodel_profiler",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["datamodel_profiler"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
